Numeric default category to avoid problems when the registry is not there

+++ src/modules/Clip/lib/Clip/Util.php
@@ -57,7 +57,7 @@ class Clip_Util
         static $id;
 
         if (!isset($id)) {
-            $id = CategoryRegistryUtil::getRegisteredModuleCategory('Clip', 'clip_pubtypes', 'Global');
+            $id = (int)CategoryRegistryUtil::getRegisteredModuleCategory('Clip', 'clip_pubtypes', 'Global');
         }
 
         return $id;
